+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Keyboard shortcut to refresh the controller (CTRL-R)
+- Button to refresh the controller (next to the Interface selection dropdown)
 
 ### Changed
 - Only displaying _Ethernet_ kind interfaces